feat: enhance input handling and remove unused scroll prevention logic
