+++ .changeset/loose-emus-grab.md
@@ -1,0 +1,5 @@
+---
+"@seed-design/react-drawer": patch
+---
+
+use-prevent-scroll 로직을 삭제합니다
