+++ packages/react-headless/drawer/src/helpers.ts
@@ -6,6 +6,27 @@ interface Style {
 
 const cache = new WeakMap();
 
+// HTML input types that do not cause the software keyboard to appear.
+const nonTextInputTypes = new Set([
+  "checkbox",
+  "radio",
+  "range",
+  "color",
+  "file",
+  "image",
+  "button",
+  "submit",
+  "reset",
+]);
+
+export function isInput(target: Element) {
+  return (
+    (target instanceof HTMLInputElement && !nonTextInputTypes.has(target.type)) ||
+    target instanceof HTMLTextAreaElement ||
+    (target instanceof HTMLElement && target.isContentEditable)
+  );
+}
+
 export function isInView(el: HTMLElement): boolean {
   const rect = el.getBoundingClientRect();
